@@ -1,6 +1,6 @@
 {
   "name": "bob-storybook",
-  "version": "1.0.0",
+  "version": "4.0.0",
   "scripts": {
     "ng": "ng",
     "start": "npm run storybook",
@@ -21,39 +21,39 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/animations": "^9.0.2",
+    "@angular/cdk": "^9.0.1",
+    "@angular/common": "^9.0.2",
+    "@angular/compiler": "^9.0.2",
+    "@angular/core": "^9.0.2",
+    "@angular/forms": "^9.0.2",
+    "@angular/material": "^9.0.1",
+    "@angular/platform-browser": "^9.0.2",
+    "@angular/platform-browser-dynamic": "^9.0.2",
+    "@angular/router": "^9.0.2",
     "@types/highcharts": "^5.0.44",
     "ag-grid-angular": "^22.1.1",
     "ag-grid-community": "^22.1.1",
     "ag-grid-enterprise": "^22.1.1",
-    "angular-froala-wysiwyg": "^3.1.0",
     "core-js": "^3.6.4",
     "date-fns": "^2.9.0",
+    "froala-editor": "^3.1.1",
     "hammerjs": "^2.0.8",
     "highcharts": "^7.1.2",
     "lodash": "^4.17.15",
-    "ng-mocks": "^8.1.0",
     "ngx-smooth-dnd": "^0.4.1",
     "rxjs": "^6.5.4",
     "tributejs": "^4.1.1",
-    "zone.js": "^0.10.2"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.23",
-    "@angular-devkit/build-ng-packagr": "^0.803.23",
-    "@angular-devkit/core": "^8.3.23",
-    "@angular/cli": "^8.3.23",
-    "@angular/compiler-cli": "^8.2.14",
-    "@angular/language-service": "^8.2.14",
+    "@angular-devkit/build-angular": "~0.900.1",
+    "@angular-devkit/build-ng-packagr": "~0.900.1",
+    "@angular-devkit/core": "^9.0.3",
+    "@angular/cli": "^9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
     "@babel/core": "^7.8.3",
     "@storybook/addon-actions": "^5.3.9",
     "@storybook/addon-knobs": "^5.3.9",
@@ -64,9 +64,9 @@
     "@storybook/storybook-deployer": "^2.8.1",
     "@types/jasmine": "^3.5.1",
     "@types/jasminewd2": "^2.0.8",
-    "@types/node": "^13.5.0",
+    "@types/node": "^12.11.1",
     "babel-loader": "^8.0.6",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^5.1.2",
     "cpx": "^1.5.0",
     "jasmine-core": "^3.5.0",
     "jasmine-marbles": "^0.6.0",
@@ -74,18 +74,17 @@
     "karma": "^4.4.1",
     "karma-chrome-launcher": "^3.1.0",
     "karma-coverage-istanbul-reporter": "^2.1.1",
-    "karma-jasmine": "^3.1.0",
-    "karma-jasmine-html-reporter": "^1.5.1",
-    "ng-packagr": "^5.7.1",
+    "karma-jasmine": "^3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "ng-packagr": "^9.0.1",
     "node-sass": "^4.13.1",
     "prettier": "^1.19.1",
     "protractor": "^5.4.2",
     "ts-node": "^8.6.2",
     "tsconfig-paths": "^3.9.0",
-    "tsickle": "^0.35.0",
-    "tslib": "^1.10.0",
     "tslint": "^6.0.0",
-    "typescript": "~3.5.3"
+    "typescript": "~3.7.5",
+    "ng-mocks": "^9.0.0"
   },
   "publishConfig": {
     "registry": "https://hibob.jfrog.io/hibob/api/npm/bob-js-virtual/"
