@@ -1,12 +1,12 @@
 {
   "name": "bob-style",
-  "version": "2.0.354",
+  "version": "4.0.0",
   "peerDependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/material": "^8.2.3",
+    "@angular/animations": "^9.0.0",
+    "@angular/cdk": "^9.0.0",
+    "@angular/common": "^9.0.0",
+    "@angular/core": "^9.0.0",
+    "@angular/material": "^9.0.0",
     "lodash": "^4.17.15",
     "hammerjs": "^2.0.8",
     "@types/node": "^13.5.0",
@@ -16,9 +16,9 @@
     "ag-grid-enterprise": "^22.1.1",
     "highcharts": "^7.1.2",
     "date-fns": "^2.9.0",
-    "angular-froala-wysiwyg": "^3.1.0",
     "tributejs": "^4.1.1",
-    "ngx-smooth-dnd": "^0.4.1"
+    "ngx-smooth-dnd": "^0.4.1",
+    "froala-editor": "^3.1.1"
   },
   "browserslist": [
     "> 1%",
