@@ -1,15 +1,24 @@
-import {AfterViewInit, ChangeDetectorRef, EventEmitter, Input, NgZone, Output} from '@angular/core';
+import {
+  AfterViewInit,
+  ChangeDetectorRef,
+  EventEmitter,
+  Input,
+  NgZone,
+  Output,
+  Directive,
+} from '@angular/core';
 import * as Highcharts from 'highcharts';
-import {Chart, ExportingMimeTypeValue, Options} from 'highcharts';
-import {ChartTypesEnum} from './chart.enum';
-import {merge} from 'lodash';
-import {simpleUID} from 'bob-style';
+import { Chart, ExportingMimeTypeValue, Options } from 'highcharts';
+import { ChartTypesEnum } from './chart.enum';
+import { merge } from 'lodash';
+import { simpleUID } from 'bob-style';
 import {
   ChartFormatterThis,
   ChartLegendAlignEnum,
   ChartLegendLayoutEnum,
   ChartLegendPositionEnum,
-  ChartLegendVerticalAlignEnum, HighChartOptions
+  ChartLegendVerticalAlignEnum,
+  HighChartOptions,
 } from './chart.interface';
 
 declare var require: any;
@@ -23,19 +32,21 @@ Boost(Highcharts);
 noData(Highcharts);
 More(Highcharts);
 
-export abstract class ChartCore implements AfterViewInit  {
+@Directive()
+export abstract class ChartCore implements AfterViewInit {
   @Input() abstract type: ChartTypesEnum;
   highChartRef: Chart;
   containerId: string = simpleUID();
   chartOptions: Options;
   options: Options;
-  private formatter = (function (component) {
-    return function () {
+  private formatter = (function(component) {
+    return function() {
       return component.tooltipFormatter(this, component);
     };
   })(this);
 
-  @Input() legendPosition: ChartLegendPositionEnum = ChartLegendPositionEnum.BOTTOM;
+  @Input() legendPosition: ChartLegendPositionEnum =
+    ChartLegendPositionEnum.BOTTOM;
   @Input() preTooltipValue = '';
   @Input() postTooltipValue = '';
   @Input() colorPalette: string[] = [
@@ -47,7 +58,8 @@ export abstract class ChartCore implements AfterViewInit  {
     '#64E572',
     '#FF9655',
     '#FFF263',
-    '#6AF9C4'];
+    '#6AF9C4',
+  ];
   @Input() height = 500;
   @Input() title: string = null;
   @Input() legend = false;
@@ -55,93 +67,96 @@ export abstract class ChartCore implements AfterViewInit  {
   @Input() pointFormat = '{series.name}: <b>{point.percentage:.1f}%</b>';
   @Input() extraOptions: HighChartOptions = {};
   @Output() legendChanged = new EventEmitter();
-  @Input() tooltipTemplate = <ChartTooltipTemplateFormatter>
-  (component: ChartCore, chartPoint: ChartFormatterThis) => `<div class="chart-tooltip">
+  @Input() tooltipTemplate = <ChartTooltipTemplateFormatter>(
+    component: ChartCore,
+    chartPoint: ChartFormatterThis
+  ) => `<div class="chart-tooltip">
       <div class="value" style="color:${chartPoint.color};">
           ${component.formatValue(chartPoint.y)}
       </div>
       <div class="key">${chartPoint.key}</div>
-    </div>`
+    </div>`;
   @Input() tooltipValueFormatter = (val: number): number | string => val;
 
-  constructor(
-    public zone: NgZone,
-    public cdr: ChangeDetectorRef
-  ) {}
+  constructor(public cdr: ChangeDetectorRef) {}
 
   tooltipFormatter(chartThis: ChartFormatterThis, component: ChartCore) {
     return this.tooltipTemplate(component, chartThis);
   }
 
   formatValue(value: number): string {
-    return `${this.preTooltipValue}${this.tooltipValueFormatter(value)}${this.postTooltipValue}`;
+    return `${this.preTooltipValue}${this.tooltipValueFormatter(value)}${
+      this.postTooltipValue
+    }`;
   }
 
   exportChart(type: ExportingMimeTypeValue) {
-    (this.highChartRef as any).exportChart(
-      {
-        type: type
-      }
-    );
+    (this.highChartRef as any).exportChart({
+      type: type,
+    });
   }
 
   initialOptions(): void {
-    this.options = merge({
-      colors: this.colorPalette,
-      chart: {
-        events: {
-          render: (event) => {
-            this.legendChanged.emit();
-          }
-        },
-        height: this.height,
-        type: this.type,
-        backgroundColor: 'rgba(255, 255, 255, 0.0)',
-        animation: {
-          duration: 200,
-        }
-      },
-      title: {
-        text: this.title,
-      },
-      legend: this.getLegendPositioning(this.legendPosition),
-      tooltip: {
-        outside: true,
-        useHTML: true,
-        style: {
-          textAlign: 'center',
-          shadow: false,
-          opacity: 1
-        },
-        formatter: this.formatter
-      },
-      plotOptions: {
-        [this.type]: {
-          animation: {},
+    this.options = merge(
+      {
+        colors: this.colorPalette,
+        chart: {
           events: {
-            afterAnimate: undefined
+            render: event => {
+              this.legendChanged.emit();
+            },
           },
-          showInLegend: this.legend,
-          dataLabels: {
-            enabled: this.showDataLabels
+          height: this.height,
+          type: this.type,
+          backgroundColor: 'rgba(255, 255, 255, 0.0)',
+          animation: {
+            duration: 200,
           },
-        }
+        },
+        title: {
+          text: this.title,
+        },
+        legend: this.getLegendPositioning(this.legendPosition),
+        tooltip: {
+          outside: true,
+          useHTML: true,
+          style: {
+            textAlign: 'center',
+            shadow: false,
+            opacity: 1,
+          },
+          formatter: this.formatter,
+        },
+        plotOptions: {
+          [this.type]: {
+            animation: {},
+            events: {
+              afterAnimate: undefined,
+            },
+            showInLegend: this.legend,
+            dataLabels: {
+              enabled: this.showDataLabels,
+            },
+          },
+        },
+        credits: {
+          enabled: false,
+        },
+        series: [],
+        exporting: {
+          enabled: false,
+        },
       },
-      credits: {
-        enabled: false
-      },
-      series: [],
-      exporting: {
-        enabled: false
-      }
-    }, this.chartOptions);
+      this.chartOptions
+    );
   }
 
   ngAfterViewInit(): void {
     this.initialOptions();
-    this.zone.runOutsideAngular(() => {
-      this.highChartRef = Highcharts.chart(this.containerId, this.options);
-    });
+    this.highChartRef = Highcharts.chart(this.containerId, this.options);
+    // this.highChartRef = Highcharts.chart(this.containerId, this.options);this.zone.runOutsideAngular(() => {
+    //   this.highChartRef = Highcharts.chart(this.containerId, this.options);
+    // });
   }
 
   applyOnChange() {
@@ -149,39 +164,43 @@ export abstract class ChartCore implements AfterViewInit  {
       this.cdr.markForCheck();
       this.initialOptions();
 
-      this.zone.runOutsideAngular(() => {
-        this.highChartRef.update(this.options, true, true);
-      });
+      this.highChartRef.update(this.options, true, true);
+      // this.zone.runOutsideAngular(() => {
+      //   this.highChartRef.update(this.options, true, true);
+      // });
     }
   }
-  private getLegendPositioning(position: ChartLegendPositionEnum, offset = {x: 0, y: 0}) {
+  private getLegendPositioning(
+    position: ChartLegendPositionEnum,
+    offset = { x: 0, y: 0 }
+  ) {
     const baseLegendPositionJson = {
       [ChartLegendPositionEnum.TOP]: {
         align: ChartLegendAlignEnum.CENTER,
         verticalAlign: ChartLegendVerticalAlignEnum.TOP,
-        layout: ChartLegendLayoutEnum.HORIZONTAL
+        layout: ChartLegendLayoutEnum.HORIZONTAL,
       },
       [ChartLegendPositionEnum.BOTTOM]: {
         align: ChartLegendAlignEnum.CENTER,
         verticalAlign: ChartLegendVerticalAlignEnum.BOTTOM,
-        layout: ChartLegendLayoutEnum.HORIZONTAL
+        layout: ChartLegendLayoutEnum.HORIZONTAL,
       },
       [ChartLegendPositionEnum.RIGHT]: {
         align: ChartLegendAlignEnum.RIGHT,
         verticalAlign: ChartLegendVerticalAlignEnum.MIDDLE,
-        layout: ChartLegendLayoutEnum.VERTICAL
+        layout: ChartLegendLayoutEnum.VERTICAL,
       },
       [ChartLegendPositionEnum.LEFT]: {
         align: ChartLegendAlignEnum.LEFT,
         verticalAlign: ChartLegendVerticalAlignEnum.MIDDLE,
-        layout: ChartLegendLayoutEnum.VERTICAL
+        layout: ChartLegendLayoutEnum.VERTICAL,
       },
     };
 
     return {
       ...baseLegendPositionJson[position],
       x: offset.x,
-      y: offset.y
+      y: offset.y,
     };
   }
 }
