@@ -1,24 +1,25 @@
-import {ChangeDetectorRef, Component, Input, NgZone, OnChanges, SimpleChanges} from '@angular/core';
-import {ChartCore} from '../chart/chart-core';
-import {SeriesLineOptions} from 'highcharts';
-import {LineChartTypes} from '../chart/chart.interface';
+import {
+  ChangeDetectorRef,
+  Component,
+  Input,
+  OnChanges,
+  SimpleChanges,
+} from '@angular/core';
+import { ChartCore } from '../chart/chart-core';
+import { SeriesLineOptions } from 'highcharts';
+import { LineChartTypes } from '../chart/chart.interface';
 
 @Component({
   selector: 'b-line-chart',
   templateUrl: '../chart/chart.component.html',
-  styleUrls: [
-    '../chart/chart.component.scss'
-  ]
+  styleUrls: ['../chart/chart.component.scss'],
 })
-export class LineChartComponent extends ChartCore implements OnChanges  {
+export class LineChartComponent extends ChartCore implements OnChanges {
   type: LineChartTypes;
   @Input() data: SeriesLineOptions[];
   @Input() name: string;
-  constructor(
-    public zone: NgZone,
-    public cdr: ChangeDetectorRef
-  ) {
-    super(zone, cdr);
+  constructor(public cdr: ChangeDetectorRef) {
+    super(cdr);
     this.height = 450;
   }
 
@@ -30,19 +31,18 @@ export class LineChartComponent extends ChartCore implements OnChanges  {
   private updateOptions() {
     this.chartOptions = {
       chart: {
-        height: Math.abs(this.height)
+        height: Math.abs(this.height),
       },
       plotOptions: {
-        [this.type]: {
-        }
+        [this.type]: {},
       },
       series: [
         {
           type: this.type as any,
           name: this.name,
-          data: this.data
-        }
-      ]
+          data: this.data,
+        },
+      ],
     };
   }
 }
