@@ -1,15 +1,4 @@
-import {
-  Input,
-  Output,
-  EventEmitter,
-  ElementRef,
-  ViewChild,
-  OnChanges,
-  SimpleChanges,
-  ChangeDetectorRef,
-  OnInit,
-  HostBinding,
-} from '@angular/core';
+import { Input, Output, EventEmitter, ElementRef, ViewChild, OnChanges, SimpleChanges, ChangeDetectorRef, OnInit, HostBinding, Directive } from '@angular/core';
 import { ButtonType, ButtonSize, BackButtonType } from './buttons.enum';
 import { Icons, IconColor, IconSize } from '../icons/icons.enum';
 import {
@@ -17,6 +6,7 @@ import {
   applyChanges,
 } from '../services/utils/functional-utils';
 
+@Directive()
 export abstract class BaseButtonElement implements OnChanges, OnInit {
   constructor(protected cd: ChangeDetectorRef) {}
 
