@@ -3,7 +3,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { EmojiChipListComponent } from './emoji-chip-list.component';
 import { MockPipe } from 'ng-mocks';
 import { EmojiFromCodePipe } from './emoji-from-code.pipe';
-import { MatTooltipModule } from '@angular/material';
+import { MatTooltipModule } from '@angular/material/tooltip';
 
 describe('EmojiChipListComponent', () => {
   let component: EmojiChipListComponent;
@@ -11,12 +11,12 @@ describe('EmojiChipListComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ MatTooltipModule ],
-      declarations: [ EmojiChipListComponent,
-        MockPipe(EmojiFromCodePipe, (val) => `${val}`)
-      ]
-    })
-    .compileComponents();
+      imports: [MatTooltipModule],
+      declarations: [
+        EmojiChipListComponent,
+        MockPipe(EmojiFromCodePipe, val => `${val}`),
+      ],
+    }).compileComponents();
   }));
 
   beforeEach(() => {
@@ -30,14 +30,17 @@ describe('EmojiChipListComponent', () => {
   });
   it('should call formatterFunction if exists and return value if it doesnt', () => {
     expect(component.valueFormatterFn(42)).toEqual(42);
-    component.valueFormatter = (val) => {
+    component.valueFormatter = val => {
       return val * 2;
     };
     expect(component.valueFormatterFn(42)).toEqual(84);
   });
   it('should call chipClick emitter when function chipClick called', () => {
     const chipClickedSpy = spyOn(component.chipClicked, 'emit');
-    component.chipClick({emoji: '1F45F', number: 245});
-    expect(chipClickedSpy).toHaveBeenCalledWith({emoji: '1F45F', number: 245});
+    component.chipClick({ emoji: '1F45F', number: 245 });
+    expect(chipClickedSpy).toHaveBeenCalledWith({
+      emoji: '1F45F',
+      number: 245,
+    });
   });
 });
