@@ -44,7 +44,7 @@ export class CheckboxComponent extends BaseFormElement implements OnChanges {
   @Input() public value = false;
   @Input() public indeterminate = false;
 
-  @Output(FormEvents.checkboxChange) changed: EventEmitter<
+  @Output('checkboxChange') changed: EventEmitter<
     InputEvent
   > = new EventEmitter<InputEvent>();
 
